feat: support defineConfig `UserConfigExport`
