+++ packages/core/src/config/constants.ts
@@ -1,0 +1,7 @@
+export const DEFAULT_CONFIG_NAMES = [
+  'farm.config.ts',
+  'farm.config.js',
+  'farm.config.mjs'
+];
+
+export const FARM_DEFAULT_NAMESPACE = 'FARM_DEFAULT_NAMESPACE';
